+++ .gitignore
@@ -5,8 +5,9 @@
 /yolo3_deepsort/deep_sort/__pycache__/
 /yolo3_deepsort/tools/__pycache__/
 /yolo3_deepsort/yolo3/__pycache__/
-/3d reconstruction/__pycache__/
-/3d reconstruction/img1/
-/3d reconstruction/img2/
-/3d reconstruction/*.ipynb
-/3d reconstruction/.ipynb_checkpoints/
+/stereo_reconstruction/__pycache__/
+/stereo_reconstruction/img1/
+/stereo_reconstruction/img2/
+/stereo_reconstruction/*.ipynb
+/stereo_reconstruction/.ipynb_checkpoints/
+*.ipynb
